feat: added grab overlay css
